Do not use pickle to output config entries in repro scripts (#100354)
New output looks like:

```
torch._dynamo.config.dynamic_shapes = True
torch._dynamo.config.assume_static_by_default = False
torch._inductor.config.fallback_random = True
torch._inductor.config.triton.cudagraphs = True
```

instead of an unreadable pickle.


Pull Request resolved: https://github.com/pytorch/pytorch/pull/100354
Approved by: https://github.com/voznesenskym

+++ torch/_dynamo/test_minifier_common.py
@@ -27,8 +27,10 @@ class MinifierTestBase(torch._dynamo.test_case.TestCase):
 
     @classmethod
     def tearDownClass(cls):
-        # Comment me out to retain temporary directory
-        shutil.rmtree(cls.DEBUG_DIR)
+        if os.getenv("PYTORCH_KEEP_TMPDIR", "0") != "1":
+            shutil.rmtree(cls.DEBUG_DIR)
+        else:
+            print("test_minifier_common tmpdir kept at: {cls.DEBUG_DIR}")
         cls._exit_stack.close()
 
     # Search for the name of the first function defined in a code string.
